+++ include/drop_role.h
