+++ include/tg_common.h
