+++ include/tsurugi_utils.h
